Updates   * Moved proxy helpers to own utils folder
