+++ repos/keg-cli/src/tasks/docker/image/run.js
@@ -4,7 +4,7 @@ const { get } = require('@keg-hub/jsutils')
 const { CONTAINERS } = require('KegConst/docker/containers')
 const { imageSelect } = require('KegUtils/docker/imageSelect')
 const { CONTAINER_PREFIXES } = require('KegConst/constants')
-const { addProxyOptions } = require('KegUtils/docker/compose/addProxyOptions')
+const { addProxyOptions } = require('KegUtils/proxy/addProxyOptions')
 const { throwDupContainerName } = require('KegUtils/error/throwDupContainerName')
 const { buildContainerContext } = require('KegUtils/builders/buildContainerContext')
 const { IMAGE } = CONTAINER_PREFIXES
