+++ repos/keg-cli/src/tasks/docker/package/run.js
@@ -2,7 +2,7 @@ const docker = require('KegDocCli')
 const { Logger } = require('KegLog')
 const { isUrl, get } = require('@keg-hub/jsutils')
 const { parsePackageUrl } = require('KegUtils/package/parsePackageUrl')
-const { addProxyOptions } = require('KegUtils/docker/compose/addProxyOptions')
+const { addProxyOptions } = require('KegUtils/proxy/addProxyOptions')
 const { buildContainerContext } = require('KegUtils/builders/buildContainerContext')
 const { CONTAINER_PREFIXES, KEG_DOCKER_EXEC, KEG_EXEC_OPTS } = require('KegConst/constants')
 const { PACKAGE } = CONTAINER_PREFIXES
