+++ repos/keg-cli/src/utils/proxy/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  ...require('./addProxyOptions'),
+  ...require('./generateComposeLabels'),
+}
